requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-multipart==0.0.9  # for form data
python-dotenv==1.0.1  # optional: for local .env loading
anthropic==0.45.2  # for Claude API integration (security scanning)
openai>=1.0.0  # for OpenRouter proxy support
